Move utf8 functions to modules/base
